Add authorization to edit & destroy methods manually

+++ app/Http/Controllers/ImageController.php
@@ -33,6 +33,10 @@ class ImageController extends Controller
     
     public function edit(Image $image)
     {
+        if (request()->user()->id !== $image->user_id) {
+            abort(403, "Access denied");
+        }
+
         return view("image.edit", compact('image'));
     }
 
@@ -44,6 +48,10 @@ class ImageController extends Controller
     
     public function destroy(Image $image)
     {
+        if (request()->user()->id !== $image->user_id) {
+            abort(403, "Access denied");
+        }
+        
         $image->delete();
         return to_route('images.index')->with('message', "Image has been removed successfully");
     }
